remove emulandice2 symbolic links that don't work in Docker
