--- modules/emulandice2/shared/emulandice_steer.R
@@ -1,1 +1,0 @@
-emulandice2/main.R
